# server/api/openapi-with-docs.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "TurbineFS API",
    "version": "v1"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Login",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Register",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [
              "Admin"
            ]
          }
        ]
      }
    },
    "/api/realtime/farm/{farmId}/listen": {
      "get": {
        "tags": [
          "RealtimeSubscriptions"
        ],
        "operationId": "RealtimeSubscriptions_ListenFarm",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeListenResponseOfFarmRealtimePayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/farm/{farmId}/turbine/{turbineId}/listen": {
      "get": {
        "tags": [
          "RealtimeSubscriptions"
        ],
        "operationId": "RealtimeSubscriptions_ListenTurbine",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "turbineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeListenResponseOfFarmRealtimePayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/farms/{farmId}/windmills": {
      "get": {
        "tags": [
          "WindmillWeb"
        ],
        "operationId": "WindmillWeb_GetWindmills",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Windmill"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/farms/{farmId}/windmills/{turbineId}/telemetry": {
      "get": {
        "tags": [
          "WindmillWeb"
        ],
        "operationId": "WindmillWeb_GetTelemetryHistory",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "turbineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2000
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WindmillTelemetry"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/farms/{farmId}/windmills/{turbineId}/alerts": {
      "get": {
        "tags": [
          "WindmillWeb"
        ],
        "operationId": "WindmillWeb_GetAlerts",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "turbineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 200
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alert"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/farms/{farmId}/windmills/{turbineId}/command": {
      "post": {
        "tags": [
          "WindmillWeb"
        ],
        "operationId": "WindmillWeb_SendCommand",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "turbineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [
              "Operator",
              "Admin"
            ]
          }
        ]
      }
    },
    "/api/farms/{farmId}/windmills/{turbineId}/actions": {
      "get": {
        "tags": [
          "WindmillWeb"
        ],
        "operationId": "WindmillWeb_GetOperatorActions",
        "parameters": [
          {
            "name": "farmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "turbineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 200
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperatorAction"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/farms/sse": {
      "get": {
        "tags": [
          "WindmillWeb"
        ],
        "operationId": "WindmillWeb_Connect",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "nickname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "nickname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "RealtimeListenResponseOfFarmRealtimePayload": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RealtimeListenResponse"
          },
          {
            "type": "object",
            "description": "Returned by subscribe endpoints with initial data. The client receives the current state\nimmediately and knows which SSE group to listen on for subsequent updates.",
            "additionalProperties": false,
            "properties": {
              "data": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FarmRealtimePayload"
                  }
                ]
              }
            }
          }
        ]
      },
      "FarmRealtimePayload": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "telemetry": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WindmillTelemetry"
            }
          },
          "alerts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Alert"
            }
          }
        }
      },
      "WindmillTelemetry": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "farmId": {
            "type": "string"
          },
          "windmillId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "windSpeed": {
            "type": "number",
            "format": "double"
          },
          "windDirection": {
            "type": "number",
            "format": "double"
          },
          "ambientTemperature": {
            "type": "number",
            "format": "double"
          },
          "rotorSpeed": {
            "type": "number",
            "format": "double"
          },
          "powerOutput": {
            "type": "number",
            "format": "double"
          },
          "nacelleDirection": {
            "type": "number",
            "format": "double"
          },
          "bladePitch": {
            "type": "number",
            "format": "double"
          },
          "generatorTemp": {
            "type": "number",
            "format": "double"
          },
          "gearboxTemp": {
            "type": "number",
            "format": "double"
          },
          "vibration": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Alert": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "farmId": {
            "type": "string"
          },
          "windmillId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "severity": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "RealtimeListenResponse": {
        "type": "object",
        "description": "Returned by subscribe endpoints so the client knows which SSE group to listen on.",
        "additionalProperties": false,
        "properties": {
          "group": {
            "type": "string"
          }
        }
      },
      "Windmill": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "farmId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      },
      "CommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string"
          },
          "payload": {}
        }
      },
      "OperatorAction": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "farmId": {
            "type": "string"
          },
          "windmillId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "operatorUserId": {
            "type": "string"
          },
          "commandType": {
            "type": "string"
          },
          "payloadJson": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}